# lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  nome: string;
  email: string;
  /** Digits only */
  telefone: string;
  /** 11 digits */
  cpf: string;
  createdAt: Date;
}