# lib/api-zod/src/generated/types/clientStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Client } from "./client";

export interface ClientStats {
  total: number;
  registeredToday: number;
  registeredThisWeek: number;
  recent: Client[];
}