# lib/api-zod/src/generated/types/createClientInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateClientInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  nome: string;
  email: string;
  /**
   * Phone number, may include formatting; digits will be extracted
   * @minLength 1
   */
  telefone: string;
  /**
   * CPF, may include formatting; must contain 11 digits
   * @minLength 1
   */
  cpf: string;
}